backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.2.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.0
openai==1.30.1
pypdf2==3.0.1
python-docx==1.1.2
python-dotenv==1.0.1
watchdog==4.0.0
aiofiles==23.2.1
beautifulsoup4==4.12.3
lxml==5.2.1
requests==2.31.0
reportlab==4.0.9
schedule==1.2.2
apscheduler==3.10.4

# Gmail email tracking (optional — only needed if you enable Gmail integration)
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.128.0
